requirements.txt
streamlit
openai
chromadb
sentence-transformers
whisper
pydub
librosa
soundfile
requests
python-dotenv
google-generativeai
